Completed (take home assignment)
I started creating this for android but decided to challenge myself and try my hand at developing the application for the iphone. I only have an old mac however and so I had to do some digging to find a legacy version of xcode to create it. It is still compatible with iphones 13, 14, and 15 however.